+++ examples/flutter_todos/lib/widgets/filter_button.dart
@@ -11,18 +11,19 @@ class FilterButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final defaultStyle = Theme.of(context).textTheme.body1;
+    final defaultStyle = Theme.of(context).textTheme.bodyText2;
     final activeStyle = Theme.of(context)
         .textTheme
-        .body1
+        .bodyText2
         .copyWith(color: Theme.of(context).accentColor);
     return BlocBuilder<FilteredTodosBloc, FilteredTodosState>(
         builder: (context, state) {
       final button = _Button(
         onSelected: (filter) {
-          BlocProvider.of<FilteredTodosBloc>(context).add(UpdateFilter(filter));
+          BlocProvider.of<FilteredTodosBloc>(context)
+              .add(FilterUpdated(filter));
         },
-        activeFilter: state is FilteredTodosLoaded
+        activeFilter: state is FilteredTodosLoadSuccess
             ? state.activeFilter
             : VisibilityFilter.all,
         activeStyle: activeStyle,