+++ packages/angular_bloc/example/example.dart
@@ -24,6 +24,7 @@ class CounterBloc extends Bloc<CounterEvent, int> {
   @override
   void onTransition(Transition<CounterEvent, int> transition) {
     print(transition);
+    super.onTransition(transition);
   }
 
   @override